+++ src/Components/Authenticate/Admin/Menu/ModalEditMenu/ModalEditMenu.jsx
@@ -5,7 +5,7 @@ import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import Modal from '@mui/material/Modal';
 import TextField from '@mui/material/TextField';
-import styles from './ModalMenu.module.css';
+import styles from '../ModalMenu/ModalMenu.module.css';
 
 const style = {
   display: 'flex',
