+++ src/Components/Authenticate/Admin/Staff/Staff.jsx
@@ -14,7 +14,7 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import MenuItem from '@mui/material/MenuItem';
 import Box from '@mui/material/Box';
-import ModalStaff from './ModalStaff';
+import ModalStaff from './ModalStaff/ModalStaff';
 import StaffBtn from './StaffBtn';
 import styles from './Staff.module.css';
 // import { flexbox } from '@mui/system';
